+++ app/models/exchanges/quote.rb
@@ -6,5 +6,7 @@ module Exchanges
 
     belongs_to :exchange
     belongs_to :securitiable, polymorphic: true
+
+    monetize :price_cents
   end
 end