+++ spec/rails_helper.rb
@@ -8,6 +8,7 @@ abort('The Rails environment is running in production mode!') if Rails.env.produ
 
 require 'spec_helper'
 require 'rspec/rails'
+require 'money-rails/test_helpers'
 
 Dir[Rails.root.join('spec/support/**/*.rb')].sort.each { |f| require f }